Changed the directive to be event driven
This was done because in the "old" version you needed to be a
sibling to the node you wanted to toggle. But now you can send a
'toggle-slider' event to the directive and it will slide up/down.

So now you call a function on the controller to send the event or
just do it like this `ng-click="$broadcast('toggle-slider')"`.

See the example for more info.